+++ src/timedecay.h
@@ -8,14 +8,14 @@
 #include <stdbool.h>
 
 typedef struct {
-	uint32_t size;          /* size of time filter */
-	uint32_t hashcount;     /* number of hashes per element */
-	uint32_t timeout;       /* number of seconds an element is valid */
-	uint32_t filter_size;   /* number of time_t values in time filter*/
+	uint64_t size;          /* size of time filter */
+	uint64_t hashcount;     /* number of hashes per element */
+	uint64_t timeout;       /* number of seconds an element is valid */
+	uint64_t filter_size;   /* number of time_t values in time filter*/
 	time_t   *filter;       /* array of time_t elements */
 } timedecay;
 
-bool timedecay_init(timedecay *, const uint32_t, const float, const uint32_t);
+bool timedecay_init(timedecay *, const uint64_t, const float, const uint64_t);
 void timedecay_destroy(timedecay);
 void timedecay_add(timedecay, const uint8_t *, const size_t);
 void timedecay_add_string(timedecay, const char *);
